Split lanczos.py into lanczos.py and slq.py
