+++ tests/test_lanczos.py
@@ -27,29 +27,6 @@ def A(n, num_significant_eigvals):
     return Q @ D @ Q.T
 
 
-@testing.parametrize("n", [200])
-@testing.parametrize("num_significant_eigvals", [30])
-@testing.parametrize("order", [10])
-# usually: ~1.5 * num_significant_eigvals.
-# But logdet seems to converge sooo much faster.
-def test_logdet(A, order):
-    n, _ = np.shape(A)
-    key = prng.PRNGKey(1)
-    received, num_nans = lanczos.trace_of_matfun(
-        np.log,
-        lambda v: A @ v,
-        order,
-        key=key,
-        num_samples_per_batch=10,
-        num_batches=1,
-        tangents_shape=(n,),
-        tangents_dtype=np.dtype(A),
-    )
-    expected = linalg.slogdet(A)[1]
-    print_if_assert_fails = ("error", np.abs(received - expected), "target:", expected)
-    assert np.allclose(received, expected, atol=1e-2, rtol=1e-2), print_if_assert_fails
-
-
 @testing.parametrize("n", [6])
 @testing.parametrize("num_significant_eigvals", [4])
 def test_tridiagonal_error_for_too_high_order(A):
